progress towards abstracting out torchdiffeq specific implementations
