+++ tests/dynamical/test_noop_interruptions.py
@@ -7,10 +7,10 @@ from chirho.dynamical.handlers import (
     DynamicInterruption,
     PointInterruption,
     PointIntervention,
-    SimulatorEventLoop,
 )
-from chirho.dynamical.ODE.handlers import simulate
-from chirho.dynamical.ops import State
+from chirho.dynamical.internals import State
+from chirho.dynamical.ODE.backends.torchdiffeq.handlers import TorchDiffEq
+from chirho.dynamical.ops import simulate
 
 from .dynamical_fixtures import UnifiedFixtureDynamics, check_trajectories_match
 
@@ -40,14 +40,14 @@ def test_noop_point_interruptions(model, init_state, tspan):
     observational_execution_result = simulate(model, init_state, tspan)
 
     # Test with standard point interruptions within timespan.
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         with PointInterruption(time=tspan[-1] / 2.0 + eps):
             result_pint = simulate(model, init_state, tspan)
 
     assert check_trajectories_match(observational_execution_result, result_pint)
 
     # Test with two standard point interruptions.
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         with PointInterruption(
             time=tspan[-1] / 4.0 + eps
         ):  # roughly 1/4 of the way through the timespan
@@ -57,7 +57,7 @@ def test_noop_point_interruptions(model, init_state, tspan):
     assert check_trajectories_match(observational_execution_result, result_double_pint1)
 
     # Test with two standard point interruptions, in a different order.
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         with PointInterruption(time=(tspan[-1] / 4.0) * 3 + eps):  # roughly 3/4
             with PointInterruption(time=tspan[-1] / 4.0 + eps):  # roughly 1/3
                 result_double_pint2 = simulate(model, init_state, tspan)
@@ -85,7 +85,7 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
     with pytest.warns(
         expected_warning=UserWarning, match="occurred after the end of the timespan"
     ):
-        with SimulatorEventLoop():
+        with TorchDiffEq():
             with PointIntervention(
                 time=post_measurement_intervention_time, intervention=intervene_state
             ):
@@ -97,7 +97,7 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
     with pytest.warns(
         expected_warning=UserWarning, match="occurred after the end of the timespan"
     ):
-        with SimulatorEventLoop():
+        with TorchDiffEq():
             with PointIntervention(
                 time=post_measurement_intervention_time, intervention=intervene_state
             ):
@@ -113,7 +113,7 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
     with pytest.warns(
         expected_warning=UserWarning, match="occurred after the end of the timespan"
     ):
-        with SimulatorEventLoop():
+        with TorchDiffEq():
             with PointIntervention(
                 time=post_measurement_intervention_time + 1.0,
                 intervention=intervene_state,
@@ -133,7 +133,7 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
 def test_point_interruption_at_start(model, init_state, tspan):
     observational_execution_result = simulate(model, init_state, tspan)
 
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         with PointInterruption(time=1.0):
             result_pint = simulate(model, init_state, tspan)
 
@@ -147,7 +147,7 @@ def test_point_interruption_at_start(model, init_state, tspan):
 def test_noop_dynamic_interruption(model, init_state, tspan, intervene_state):
     observational_execution_result = simulate(model, init_state, tspan)
 
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         tt = (tspan[-1] - tspan[0]) / 2.0
         with DynamicInterruption(
             event_f=lambda t, _: torch.where(t < tt, tt - t, 0.0),
