+++ tests/dynamical/test_static_observation.py
@@ -9,10 +9,10 @@ from pyro.infer.autoguide import AutoMultivariateNormal
 from chirho.dynamical.handlers import (
     NonInterruptingPointObservationArray,
     PointObservation,
-    SimulatorEventLoop,
 )
-from chirho.dynamical.ODE.handlers import simulate
-from chirho.dynamical.ops import State
+from chirho.dynamical.internals import State
+from chirho.dynamical.ODE.backends.torchdiffeq.handlers import TorchDiffEq
+from chirho.dynamical.ops import simulate
 
 from .dynamical_fixtures import (
     UnifiedFixtureDynamics,
@@ -54,7 +54,7 @@ def test_multiple_point_observations(model):
     S_obs = torch.tensor(10.0)
     data1 = {"S_obs": S_obs}
     data2 = {"I_obs": torch.tensor(5.0), "R_obs": torch.tensor(5.0)}
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         with PointObservation(time=3.1, data=data2):
             with PointObservation(time=2.9, data=data1):
                 result = simulate(model, init_state, tspan)
@@ -90,7 +90,7 @@ def test_log_prob_exists(model, obs_handler):
     S_obs = torch.tensor(10.0)
     data = {"S_obs": S_obs}
     with pyro.poutine.trace() as tr:
-        with SimulatorEventLoop():
+        with TorchDiffEq():
             with _get_compatible_observations(obs_handler, time=2.9, data=data):
                 simulate(model, init_state, tspan)
 
@@ -108,7 +108,7 @@ def test_tspan_collision(model, obs_handler):
     """
     S_obs = torch.tensor(10.0)
     data = {"S_obs": S_obs}
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         with _get_compatible_observations(obs_handler, time=tspan[1], data=data):
             result = simulate(model, init_state, tspan)
 
@@ -131,7 +131,7 @@ def test_svi_composition_test_one(model, obs_handler):
     class ConditionedSIR(pyro.nn.PyroModule):
         def forward(self):
             sir = model()
-            with SimulatorEventLoop():
+            with TorchDiffEq():
                 with _get_compatible_observations(obs_handler, time=2.9, data=data1):
                     traj = simulate(sir, init_state, tspan)
             return traj
@@ -156,7 +156,7 @@ def test_interrupting_and_non_interrupting_observation_array_equivalence(model):
     times = torch.tensor([1.5, 2.9, 3.2])
 
     with pyro.poutine.trace() as tr1:
-        with SimulatorEventLoop():
+        with TorchDiffEq():
             with PointObservation(
                 time=times[1].item(), data={k: v[1] for k, v in data.items()}
             ):
@@ -169,7 +169,7 @@ def test_interrupting_and_non_interrupting_observation_array_equivalence(model):
                         interrupting_ret = simulate(model, init_state, tspan)
 
     with pyro.poutine.trace() as tr2:
-        with SimulatorEventLoop():
+        with TorchDiffEq():
             with NonInterruptingPointObservationArray(times=times, data=data):
                 non_interrupting_ret = simulate(model, init_state, tspan)
 
@@ -187,7 +187,7 @@ def test_point_observation_at_tspan_start_excepts(model, init_state, tspan):
     This occurs right now due to an undiagnosed error, so this test is a stand-in until that can be fixed.
     """
 
-    with SimulatorEventLoop():
+    with TorchDiffEq():
         with pytest.raises(ValueError, match="occurred at the start of the timespan"):
             with PointObservation(time=tspan[0], data={"S_obs": torch.tensor(10.0)}):
                 simulate(model, init_state, tspan)
@@ -218,7 +218,7 @@ def test_svi_composition_test_multi_point_obs(model):
                 obs_time = obs[0].item()
                 obs_data = obs[1]
                 observation_managers.append(PointObservation(obs_time, obs_data))
-            with SimulatorEventLoop():
+            with TorchDiffEq():
                 with ExitStack() as stack:
                     for manager in observation_managers:
                         stack.enter_context(manager)
@@ -244,7 +244,7 @@ def test_svi_composition_vectorized_obs(model):
     class ConditionedSIR(pyro.nn.PyroModule):
         def forward(self):
             sir = model()
-            with SimulatorEventLoop():
+            with TorchDiffEq():
                 with NonInterruptingPointObservationArray(times=times, data=data):
                     traj = simulate(sir, init_state, tspan)
             return traj
@@ -275,7 +275,7 @@ def test_simulate_persistent_pyrosample(use_event_loop):
         S_obs = torch.tensor(10.0)
         data1 = {"S_obs": S_obs}
         data2 = {"I_obs": torch.tensor(5.0), "R_obs": torch.tensor(5.0)}
-        with SimulatorEventLoop():
+        with TorchDiffEq():
             with PointObservation(time=3.1, data=data2):
                 with PointObservation(time=2.9, data=data1):
                     result = simulate(model, init_state, tspan)
